[junu0516/airbnb#24] Chore: 파일이름 변경 Accomodation -> Room
