+++ iOS/Airbnb/ExtraSearch/ConditionSettingViewController.swift
@@ -69,7 +69,7 @@ final class ConditionSettingViewController: UIViewController {
     
     //추후 검색 결과 리스트 보여주는 화면으로 넘어가는 로직으로 대체해야 함
     @objc private func pushNextViewController() {
-        self.navigationController?.pushViewController(AccomodationListViewController(), animated: true)
+        self.navigationController?.pushViewController(RoomListViewController(), animated: true)
     }
 }
