+++ iOS/Airbnb/Server/ContentType.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+enum ContentType {
+    case json
+    case image
+    var value: String {
+        switch self {
+        case .json:
+            return "application/json"
+        case .image:
+            return "image/png"
+        }
+    }
+}
