+++ iOS/Airbnb/Server/HttpMethod.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum HttpMethod {
+    case get
+    case post
+}
